requirements.txt
openai
python-dotenv
streamlit==1.33.0
moviepy
